refactor: move data sample to assets/ data
