+++ src/screens/OrderDetailScreen.tsx
@@ -27,6 +27,7 @@ import InputReviewArea from "../components/orderDetail/InputReviewArea";
 import SolarPenBold from "../../assets/images/icons/SolarPenBold";
 import GradientView from "../components/gradientView/GradientView";
 import SolarBag5Bold from "../../assets/images/icons/SolarBag5Bold";
+import {orders} from "../../assets/data/order/order";
 
 type OrderDetailScreenProps = {
 	route: RouteProp<RootStackParamList, "OrderDetailScreen">;
@@ -245,137 +246,3 @@ const makeStyled = (theme: ThemeType) =>
 
 	})
 
-// Khởi tạo dữ liệu cho các phần tử
-const orders: OrderDetailType[] = [
-	{
-		id: 'SP1',
-		price: 1000,
-		starReview: 5,
-		status: ORDER_STATUS_COMPLETED,
-		products: [
-			{
-				id: 'prod1',
-				image: burrito,
-				name: 'Apple',
-				rating: 5,
-				price: 60000,
-				discountInfo: {
-					discount: 0.5,
-					expired: new Date("2027-12-31"),
-				},
-				description: 'Fresh red apples',
-				options: [
-					{id: 'opt1', name: 'No Sugar', price: 0.3},
-					{
-						id: 'grp1',
-						name: 'Chan Chau',
-						options: [
-							{id: 'opt2', name: 'den', price: 0.5},
-							{id: 'opt3', name: 'trang', price: 0},
-						],
-					},
-				],
-			},
-			{
-				id: 'prod12',
-				image: banh_mi,
-				name: 'Banh mi',
-				rating: 4,
-				price: 50000,
-				discountInfo: {
-					discount: 0.5,
-					expired: new Date("2027-12-31"),
-				},
-				description: '',
-			},
-		],
-		address: {
-			name: 'Home',
-			address: '123 Main St',
-		},
-		paymentMethod: {type: 'Credit Card'},
-		promotions: [{name: 'PROMO10'}, {name: 'PREE'}],
-		subTotal: 4.5,
-		deliveryFee: 2.0,
-		discount: 0.5,
-		description: '',
-	},
-	{
-		id: 'SP2',
-		price: 2000,
-		starReview: 4,
-		status: ORDER_STATUS_CANCELLED,
-		products: [
-			{
-				id: 'prod2',
-				image: burger,
-				name: 'Banana',
-				rating: 4,
-				discountInfo: {
-					discount: 0.5,
-					expired: new Date("2027-12-31"),
-				},
-				price: 0.75,
-				description: 'Fresh yellow bananas',
-				options: [{id: 'opt4', name: 'Organic', price: 1}],
-			},
-			{
-				id: 'prod22',
-				image: donut_1,
-				name: 'Banana',
-				rating: 4,
-				discountInfo: {
-					discount: 0.5,
-					expired: new Date("2027-12-31"),
-				},
-				price: 0.75,
-				description: 'Fresh yellow bananas',
-				options: [{id: 'opt4', name: 'Organic', price: 1}],
-			},
-		],
-		address: {
-			name: 'Home',
-			address: '123 Main St',
-		},
-		paymentMethod: {type: 'Credit Card'},
-		promotions: [{name: 'PROMO10'}, {name: 'PREE DELIVERY'}],
-		subTotal: 5.5,
-		deliveryFee: 1.5,
-		discount: 1.0,
-		description: 'This is a very good product',
-	},
-	{
-		id: 'SP3',
-		price: 3000,
-		starReview: 3,
-		status: ORDER_STATUS_ACTIVE,
-		products: [
-			{
-				id: 'prod3',
-				name: 'Orange Juice',
-				image: lemonade,
-				rating: 0,
-				price: 3.0,
-				description: 'Freshly squeezed orange juice',
-			},
-			{
-				id: 'prod31',
-				image: pasta,
-				rating: 0,
-				name: 'Orange Juice',
-				price: 3.0,
-				description: 'Freshly squeezed orange juice',
-			},
-		],
-		address: {
-			name: 'Home',
-			address: '123 Main St',
-		},
-		paymentMethod: {type: 'Credit Card'},
-		promotions: [{name: 'PROMO10'}, {name: 'PREE DELIVERY'}],
-		subTotal: 6.0,
-		deliveryFee: 1.0,
-		discount: 0.5,
-		description: 'This is a very good product',
-	},
-];
